--- fhelp-core/src/main/java/ru/bukharov/fhelp/dao/AccountDAO.java
@@ -1,7 +1,0 @@
-package ru.bukharov.fhelp.dao;
-
-import org.springframework.data.repository.CrudRepository;
-import ru.bukharov.fhelp.domain.Account;
-
-public interface AccountDAO extends CrudRepository<Account, Long> {
-}
