+++ fhelp-core/src/main/java/ru/bukharov/fhelp/dao/RepositoryConfig.java
@@ -3,8 +3,8 @@ package ru.bukharov.fhelp.dao;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.rest.core.config.RepositoryRestConfiguration;
 import org.springframework.data.rest.webmvc.config.RepositoryRestConfigurerAdapter;
-import ru.bukharov.fhelp.domain.Account;
-import ru.bukharov.fhelp.domain.AccountState;
+import ru.bukharov.fhelp.domain.account.Account;
+import ru.bukharov.fhelp.domain.account.AccountState;
 
 @Configuration
 public class RepositoryConfig extends RepositoryRestConfigurerAdapter {
