+++ fhelp-core/src/main/java/ru/bukharov/fhelp/dao/account/AccountStateDAO.java
@@ -1,7 +1,7 @@
-package ru.bukharov.fhelp.dao;
+package ru.bukharov.fhelp.dao.account;
 
 import org.springframework.data.repository.CrudRepository;
-import ru.bukharov.fhelp.domain.AccountState;
+import ru.bukharov.fhelp.domain.account.AccountState;
 
 public interface AccountStateDAO extends CrudRepository<AccountState, Long> {
 }
