+++ fhelp-core/src/main/java/ru/bukharov/fhelp/domain/account/AccountState.java
@@ -1,16 +1,13 @@
-package ru.bukharov.fhelp.domain;
+package ru.bukharov.fhelp.domain.account;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
 import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
-import javax.persistence.MappedSuperclass;
 import javax.persistence.Table;
 import java.util.Date;
