+++ fhelp-core/src/main/java/ru/bukharov/fhelp/service/account/AccountServiceException.java
@@ -1,0 +1,8 @@
+package ru.bukharov.fhelp.service.account;
+
+public class AccountServiceException extends Exception {
+
+    public AccountServiceException(String message) {
+        super(message);
+    }
+}
