+++ fhelp-core/src/main/java/ru/bukharov/fhelp/service/user/UserService.java
@@ -1,6 +1,7 @@
 package ru.bukharov.fhelp.service.user;
 
 import org.springframework.security.core.userdetails.UserDetailsService;
+import ru.bukharov.fhelp.domain.user.User;
 import ru.bukharov.fhelp.dto.user.UserDTO;
 
 public interface UserService extends UserDetailsService {
@@ -8,4 +9,6 @@ public interface UserService extends UserDetailsService {
     UserDTO saveUser(UserDTO userDTO);
 
     String getCurrentLoggedInUsername();
+
+    User getCurrentLoggedInUser();
 }
